# Optimize camel-casing logic