--- a/src/NJsonSchema.Benchmark/Program.cs
+++ b/src/NJsonSchema.Benchmark/Program.cs
@@ -7,7 +7,7 @@ namespace NJsonSchema.Benchmark
         public static void Main(string[] args)
         {
             // RunCsharpBenchmark();
-            BenchmarkDotNet.Running.BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).RunAllJoined();
+            BenchmarkDotNet.Running.BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run();
         }
 
 #pragma warning disable IDE0051